requirements.txt
Django
djangorestframework
lxml
numpy
pandas
requests